refactor(comments): share PR comment core
