+++ src/renderer/src/components/pr-comments-resolution-prompt.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import type { PRComment } from '../../../shared/types'
-import { groupPRComments } from '@/lib/pr-comment-groups'
+import { groupPRComments } from '../../../shared/pr-comment-groups'
 import {
   buildPRCommentsResolutionPrompt,
   isResolvablePRCommentGroup
