--- src/renderer/src/components/right-sidebar/pr-comment-filters.ts
@@ -1,5 +1,0 @@
-export {
-  filterPRCommentsByAudience,
-  isBotPRComment as isAutomatedPRComment,
-  type PRCommentAudienceFilter
-} from '@/lib/pr-comment-audience'
