+++ src/renderer/src/components/right-sidebar/pr-comments-ai-launch-ack.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest'
 import type { PRComment } from '../../../../shared/types'
-import type { PRCommentGroup } from '@/lib/pr-comment-groups'
+import type { PRCommentGroup } from '../../../../shared/pr-comment-groups'
 import {
   acknowledgePRCommentsAfterAiLaunch,
   attachPRReviewReplyParent,
