+++ src/renderer/src/lib/pr-comment-action-state.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import type { PRComment } from '../../../shared/types'
-import { groupPRComments } from './pr-comment-groups'
+import { groupPRComments } from '../../../shared/pr-comment-groups'
 import {
   getPRCommentGroupActionState,
   isPRCommentGroupQueueableForAI,
