+++ src/renderer/src/lib/pr-comment-resolution-classes.ts
@@ -1,0 +1,3 @@
+export const PR_COMMENT_RESOLVED_CONTAINER_CLASS = 'opacity-50'
+export const PR_COMMENT_RESOLVED_AUTHOR_CLASS = 'text-muted-foreground'
+export const PR_COMMENT_OPEN_AUTHOR_CLASS = 'text-foreground'
