+++ src/shared/pr-comment-audience.ts
@@ -1,35 +1,19 @@
-import type { PRComment } from '../../../../src/shared/types'
-import {
-  createBotAuthorOverrideSet,
-  normalizePRCommentAuthorLogin
-} from '../../../../src/shared/pr-bot-author-overrides'
+import type { PRComment } from './types'
+import { normalizePRCommentAuthorLogin } from './pr-bot-author-overrides'
 
-export { createBotAuthorOverrideSet }
-
-// Audience filtering for the PR comment timeline, ported from the desktop helper
-// (src/renderer/src/lib/pr-comment-audience.ts) minus its i18n wrapper so it stays
-// pure + unit-testable under the node Vitest config. Classification must match the
-// desktop so the same comment reads as human/bot on both surfaces.
 export type PRCommentAudienceFilter = 'all' | 'human' | 'bot'
-
-export const PR_COMMENT_AUDIENCE_FILTERS: { value: PRCommentAudienceFilter; label: string }[] = [
-  { value: 'all', label: 'All' },
-  { value: 'human', label: 'Humans' },
-  { value: 'bot', label: 'Bots' }
-]
 
 const BOT_LOGIN_SUFFIX = '[bot]'
 const AUTOMATION_LOGIN_PATTERNS = [
   /bot$/i,
-  /-bot$/i,
   /\bbot\b/i,
   /automation/i,
   /actions/i,
   /renovate/i,
   /dependabot/i
 ]
-// Some AI/code-review services use regular user accounts, so GitHub metadata can
-// report them as users — keep this list in sync with the desktop helper.
+
+// GitHub can report regular-account review services as users.
 const KNOWN_AUTOMATION_LOGIN_SUBSTRINGS = [
   'chatgpt-codex-connector',
   'codex-connector',
@@ -55,10 +39,7 @@ export function isBotPRComment(
 ): boolean {
   const author = comment.author.trim()
   const normalized = normalizePRCommentAuthorLogin(author)
-  if (botAuthorOverrides?.has(normalized)) {
-    return true
-  }
-  if (comment.isBot === true) {
+  if (botAuthorOverrides?.has(normalized) || comment.isBot === true) {
     return true
   }
   if (normalized.endsWith(BOT_LOGIN_SUFFIX)) {
@@ -71,15 +52,11 @@ export function isBotPRComment(
 }
 
 export function getPRCommentAudienceCounts(
-  comments: PRComment[],
+  comments: readonly PRComment[],
   botAuthorOverrides?: ReadonlySet<string>
 ): Record<PRCommentAudienceFilter, number> {
   const bot = comments.filter((comment) => isBotPRComment(comment, botAuthorOverrides)).length
-  return {
-    all: comments.length,
-    human: comments.length - bot,
-    bot
-  }
+  return { all: comments.length, human: comments.length - bot, bot }
 }
 
 export function filterPRCommentsByAudience(
@@ -94,15 +71,4 @@ export function filterPRCommentsByAudience(
     return comments.filter((comment) => !isBotPRComment(comment, botAuthorOverrides))
   }
   return comments
-}
-
-export function getPRCommentAudienceEmptyLabel(filter: PRCommentAudienceFilter): string {
-  switch (filter) {
-    case 'bot':
-      return 'No bot comments.'
-    case 'human':
-      return 'No human comments.'
-    case 'all':
-      return 'No comments yet.'
-  }
 }
